Fix identity statement view dependency cycle
