+++ components/user/identity/statements/add/UserPageIdentityAddStatements.constants.ts
@@ -1,0 +1,7 @@
+export enum STATEMENT_ADD_VIEW {
+  SELECT = "SELECT",
+  CONTACT = "CONTACT",
+  NFT_ACCOUNT = "NFT_ACCOUNT",
+  SOCIAL_MEDIA_ACCOUNT = "SOCIAL_MEDIA_ACCOUNT",
+  SOCIAL_MEDIA_VERIFICATION_POST = "SOCIAL_MEDIA_VERIFICATION_POST",
+}
